Added BaseFronteraScheduler so schedulers can automatically add required middlewares
