+++ frontera/contrib/scrapy/schedulers/frontier.py
@@ -1,20 +1,15 @@
 from collections import deque
 
-from collections import deque
 from time import time
 
-from scrapy.core.scheduler import Scheduler
 from scrapy.http import Request
 from scrapy import log
-from scrapy.utils.misc import load_object
 
+from frontera.contrib.scrapy.schedulers import BaseFronteraScheduler
 from frontera.contrib.scrapy.manager import ScrapyFrontierManager
 from frontera.settings import Settings
 
 STATS_PREFIX = 'frontera'
-
-DOWNLOADER_MIDDLEWARE = 'frontera.contrib.scrapy.middlewares.schedulers.SchedulerDownloaderMiddleware'
-SPIDER_MIDDLEWARE = 'frontera.contrib.scrapy.middlewares.schedulers.SchedulerSpiderMiddleware'
 
 
 class StatsManager(object):
@@ -75,10 +70,9 @@ class StatsManager(object):
         self.stats.set_value(self._get_stats_name(variable), value)
 
 
-class FronteraScheduler(Scheduler):
+class FronteraScheduler(BaseFronteraScheduler):
 
     def __init__(self, crawler):
-
         # Add scrapy integration middlewares for scheduler
         self._add_middlewares(crawler)
 
@@ -194,22 +188,3 @@ class FronteraScheduler(Scheduler):
             if overused_factor > self.frontier.manager.settings.get('OVERUSED_SLOT_FACTOR'):
                 info['overused_keys'].append(key)
         return info
-
-    def _add_middlewares(self, crawler):
-        """
-        Adds crawl-frontier scrapy scheduler downloader and spider middlewares.
-        Hack to avoid defining crawl-frontier scrapy middlewares in settings.
-        Middleware managers (downloader+spider) has already been initialized at this moment.
-        """
-        self._add_middleware_to_manager(manager=crawler.engine.downloader.middleware,
-                                        mw=load_object(DOWNLOADER_MIDDLEWARE).from_crawler(crawler))
-        self._add_middleware_to_manager(manager=crawler.engine.scraper.spidermw,
-                                        mw=load_object(SPIDER_MIDDLEWARE).from_crawler(crawler))
-
-    def _add_middleware_to_manager(self, manager, mw):
-        """
-        Adds mw to already initialized middleware manager.
-        Reproduces the mw add process at the end of the middleware manager mws list.
-        """
-        manager.middlewares = manager.middlewares + (mw,)
-        manager._add_middleware(mw)
